+++ integration-tests/emulated/tests/assets/asset-hub-polkadot/src/tests/mod.rs
@@ -22,3 +22,4 @@ mod set_xcm_versions;
 mod swap;
 mod teleport;
 mod treasury;
+mod xcm_fee_estimation;